refactor: move intro catch-up into whatsapp platform
