+++ src/platforms/whatsapp/runtime.ts
@@ -2,7 +2,7 @@ import type { WASocket } from '@whiskeysockets/baileys';
 import { logger } from '../../middleware/logger.js';
 import { startConnection } from './connection.js';
 import { registerWhatsAppHandlers } from './handlers.js';
-import { registerIntroCatchUp } from '../../features/introductions.js';
+import { registerIntroCatchUp } from './introductions-catchup.js';
 import { scheduleDigest } from '../../features/digest.js';
 import type { PlatformRuntime } from '../types.js';
